pyproject.toml
[project]
name = "orbidi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "dbt-bigquery>=1.11.0",
    "functions-framework>=3.10.0",
    "google-cloud-bigquery>=3.39.0",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-storage>=3.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
